third_party/WebKit/LayoutTests/http/tests/htmlimports/cors-same-origin.html

<!DOCTYPE html>
<html>
<head>
<link id="basic" rel="import" href="http://localhost:8080/htmlimports/resources/cors-basic.cgi">
<link id="nested" rel="import" href="http://localhost:8080/htmlimports/resources/cors-subimports.cgi">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>

shouldBeEqualToString("basic.import.querySelector('h1').innerHTML", "Hello, CORS!");
shouldBeNull("nested.import.querySelector('#sameOriginNoCors').import");
shouldBeEqualToString("nested.import.querySelector('#sameOriginCors').import.querySelector('h1').innerHTML", "Hello, CORS!");
shouldBeEqualToString("nested.import.querySelector('#masterOriginNoCors').import.querySelector('h1').innerHTML", "Hello");
shouldBeEqualToString("nested.import.querySelector('#masterOriginCors').import.querySelector('h1').innerHTML", "Hello, CORS!");

</script>
</body>
</html>
